Restore warnings with no pause and incremental pauses
Make it so quiet days defaults to zero for negative numbers so we're not breaking existing clients. When days old reaches the quiet days (meaning, it's actually *over* by one day, since days old starts at zero, so, if quiet days == 1 and days old == 1, it means the quiet day was day=0 of the package), we start issuing warnings, but the pause may still be zero (as in the previous example, 1-1).

In that case, warn will be true, but pause will be 0, so we need to also return an empty suffix, since otherwise we get a weird message saying the pause was zero :).

We also reintroduce the increasing pauses after quiet days have passed, going from very short pauses (down to zero the first day) up to the max pause configured. This should make the transition smoother once warnings start appearing, instead of causing a panic run for an alternative nuget :)