Make sure transitive targets are always properly packed
For non-development dependencies, analyzers will flow transitively.
This means that there is a need for transitive targets that also flow.

Rather than trying to fit this in the authoring package (which isn't trivial since the sponsorable package can itself have its own targets, transitive or not, and would need to properly import these), create a separate package that provides the build side of things.

This will only be necessary as long as the transitiveness of analyzers is an unresolved issue on the nuget side. Eventually, we might just be able to embed this targets file within each sponsorable package.