Fix transitive check bug that caused no checks at all
Since we were not emitting the package dependency items, the transitive check was always empty. In combination with a bug that considered an empty string (parent package) value as a signal to skip, no checks were being performed at all.

This only manifested if NuGetizer wasn't installed, since it already turns on the dependency item generation too for its own purposes.