Delete entire analyzer/package project which is no longer needed
Now that the gh-sponsors CLI provides a standards-based mechanism to check for sponsorships (a JWT manifest that can be read using standard libraries), we don't need any of the existing code.

This makes for a far lower barrier of adoption, removes the need for SL to make decisions for integrators and shifts the discussion to one about the value of knowing who is sponsoring a project for the purpose of feature-toggling functionality.