Add a development-only and source-only helper package
The Devlooped.SponsorLink package is now nothing but a helper to simplify integrating the SL manifest checks into an analyzer, build task, or some other .NET integration point.

It provides an assembly-level `[Funding(account)]` attribute to be used for automatically checking one or more sponsorable accounts that the user can choose to fund for a given project. This is populated automatically from the .github/FUNDING.yml if present. It can be extended in two ways:

* MSBuild: add `<GitHubSponsorable Include="[account]" />` items
* Add `[assembly: Funding("account")]` manually

The static SponsorLink class is imported directly from the gh-sponsors repo via `dotnet file add/sync` so we can keep it in sync with that repo, but should otherwise be low-churn.

The Helpers.cs itself offers a facade with more integration-side APIs, such as checking for an active editor being used, retrieving a user's email from a working directory (for local offline manifest checks), etc.