Add a delete endpoint for data deletion if needed
This must exist and be invokable by the CLI since it will call backends from multiple sponsorable accounts eventually. Some other vendors might decide to persist user data, so this is a hook for them to delete it when requested by the user.

We don't plan on persisting anything server-side, but this is for completeness and future proofing.