Add tier retrieval and metadata population
Tiers are chained logically by GH sponsors API by exposing what's the previous tier to any particular one (except for the lowest-level first tier).

Based on this and the fact that you can "hide" metadata in item descriptions throughout GH (i.e. in issues and PRs) via HTML comments like `<!-- key: value -->`, we populate metadata for each tier which will later on become claims in the JWT for the user, which at run-time can be checked by a sponsorable library.

Features would be enabled/disabled based on tier metadata. In fact, this metadata could even be used as a feature flags mechanism entirely.