Improve development authentication callback from GH app
A new app can be provided that points to an ngrok-forwarded domain. This new code will replicate as closeely as possible the full EasyAuth in production, includig the /me endpoint and its claims.

Combined with the device flow autentication (also used in local dev only), which reads the new `access_token` cookie, and performs the proper authentication for API requests from the browser.