When auth fails during nuget stats, render failure
It can be helpful to see the actual error from the GH CLI when authentication fails. We only propagate the error as the output to the caller from our process helper if there was no output already (i.e. some graphql APIs return failure but still provide data). This eliminates the chance of breaking existing code that relied on that.

Also added a debug-only option to pass in a direct token, which is easier for debug runs of the CLI.