Account for source repo name/ownership changes for nuget stats
Packages may have specified a certain owner/repo information, which can later be changed on github. Repositories can be renamed, and ownership transferred, which would now leave an inconsistency in how we report since the current info on github wouldn't match a published package.

We account for this by first resolving the full name of the owner/repo we find in package metadata, which would account for any renaming that happened.