Fix overwriting of $(FundingPackageId) in targets
The existig Devlooped.Sponsors.targets had a bug in that it used the same property name the analyzer uses, for its temporary projection from @(FundingPackageId). This caused the property to overwritten, which manifested as soon as we bumped ThisAssembly.AssemblyInfo, which imports the buggy targets into the sample analyzer too now.

By using `_FundingPackage` as the property name instead, analyzers leveraging ThisAssembly won't have an issue from now on, and more importantly: we remove the limitation that only non-SLv2 versions of ThisAssembly could be used previously (unless you implemented an unintuitive workaround like we do in this commit in the Analyzer.csproj via the @(FundingPackageId) item, which won't be necessary from now on (as soon as we can bump again ThisAssembly).