Create non-draft releases with updated body for draft releases
Turns out that editing a release body does not nofity mentioned sponsors when the edit happens. So we introduce a new conventional approach: we will create releases as draft, that will invoke the hook and cause the draft to be deleted entirely and recreated as a non-draft with the new updated body. According to a Grok search, this will properly notify mentioned users, and it's apparently by design to avoid spamming mentioned users on multiple edits of the same release (which sort of makes sense).